refactor(retrieval): extract RetrievalResult to models and add lazy imports
